refactor: 상품 상세 화면 코드 정리
